workspace bug fixed, UI update

+++ BMPEditor/Editor/workspace.cpp
@@ -300,17 +300,17 @@ void Workspace::paintEvent(QPaintEvent *event) {
 
 
         // vekresleni meritek ose
-        int ruleXStart = 0;
+        float ruleXStart = 0;
         while(ruleXStart > -offset.x() && this->scale != 0) {
             ruleXStart -= this->scale * 4;
         }
-        int ruleYStart = 0;
+        float ruleYStart = 0;
         while(ruleYStart > -offset.y() && this->scale != 0) {
             ruleYStart -= this->scale * 4;
         }
 
-        int ruleXEnd = this->width() - offset.x();
-        int ruleYEnd = this->height() - offset.y();
+        float ruleXEnd = this->width() - offset.x();
+        float ruleYEnd = this->height() - offset.y();
 
         // x osa meritko
         float rulePos;
